+++ packages/video_player/video_player_android/android/src/main/java/io/flutter/plugins/videoplayer/texture/TextureVideoPlayer.java
@@ -12,6 +12,7 @@ import androidx.annotation.RestrictTo;
 import androidx.annotation.VisibleForTesting;
 import androidx.media3.common.MediaItem;
 import androidx.media3.common.util.UnstableApi;
+import androidx.media3.exoplayer.DefaultRenderersFactory;
 import androidx.media3.exoplayer.ExoPlayer;
 import io.flutter.plugins.videoplayer.ExoPlayerEventListener;
 import io.flutter.plugins.videoplayer.VideoAsset;
@@ -55,15 +56,26 @@ public final class TextureVideoPlayer extends VideoPlayer implements SurfaceProd
         surfaceProducer,
         asset.getMediaItem(),
         options,
-        () -> {
-          androidx.media3.exoplayer.trackselection.DefaultTrackSelector trackSelector =
-              new androidx.media3.exoplayer.trackselection.DefaultTrackSelector(context);
-          ExoPlayer.Builder builder =
-              new ExoPlayer.Builder(context)
-                  .setTrackSelector(trackSelector)
-                  .setMediaSourceFactory(asset.getMediaSourceFactory(context));
-          return builder.build();
-        });
+        () -> createExoPlayer(context, asset, options, new DefaultRenderersFactory(context)));
+  }
+
+  // TODO: Migrate to stable API, see https://github.com/flutter/flutter/issues/147039.
+  @UnstableApi
+  @VisibleForTesting
+  @NonNull
+  public static ExoPlayer createExoPlayer(
+      @NonNull Context context,
+      @NonNull VideoAsset asset,
+      @NonNull VideoPlayerOptions options,
+      @NonNull DefaultRenderersFactory renderersFactory) {
+    androidx.media3.exoplayer.trackselection.DefaultTrackSelector trackSelector =
+        new androidx.media3.exoplayer.trackselection.DefaultTrackSelector(context);
+    renderersFactory.setEnableDecoderFallback(options.enableDecoderFallback);
+    ExoPlayer.Builder builder =
+        new ExoPlayer.Builder(context, renderersFactory)
+            .setTrackSelector(trackSelector)
+            .setMediaSourceFactory(asset.getMediaSourceFactory(context));
+    return builder.build();
   }
 
   // TODO: Migrate to stable API, see https://github.com/flutter/flutter/issues/147039.