+++ packages/video_player/video_player/CHANGELOG.md
@@ -1,5 +1,8 @@
-## NEXT
+## 2.12.0
 
+* Adds `VideoPlayerAndroidOptions.enableDecoderFallback` to allow Android's
+  ExoPlayer implementation to fall back to another decoder if decoder
+  initialization fails.
 * Updates minimum supported SDK version to Flutter 3.38/Dart 3.10.
 
 ## 2.11.1